requirements.txt
streamlit==1.29.0
plotly==5.18.0
pandas==2.0.3
numpy==1.24.4
folium==0.15.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.17.3
Pillow==10.1.0